Reverted to parsing scale only in EOS
